--- src/main/java/com/thevale/moretimecapsulesmod/proxy/ClientProxy.java
@@ -1,12 +1,0 @@
-package com.thevale.moretimecapsulesmod.proxy;
-
-import net.minecraft.client.Minecraft;
-import net.minecraft.world.World;
-
-public class ClientProxy implements IProxy{
-
-    @Override
-    public World getClientWorld() {
-        return Minecraft.getInstance().world;
-    }
-}
