--- src/main/java/com/thevale/moretimecapsulesmod/proxy/IProxy.java
@@ -1,8 +1,0 @@
-package com.thevale.moretimecapsulesmod.proxy;
-
-import net.minecraft.world.World;
-
-public interface IProxy {
-
-    World getClientWorld();
-}
