--- src/main/java/com/thevale/moretimecapsulesmod/proxy/ServerProxy.java
@@ -1,11 +1,0 @@
-package com.thevale.moretimecapsulesmod.proxy;
-
-import net.minecraft.world.World;
-
-public class ServerProxy implements IProxy {
-
-    @Override
-    public World getClientWorld() {
-        throw new IllegalStateException("Only Run this on the Client");
-    }
-}
