+++ src/main/resources/META-INF/mods.toml
@@ -15,7 +15,7 @@ license="All rights reserved"
 # The modid of the mod
 modId="moretimecapsulesmod" #mandatory
 # The version number of the mod - there's a few well known ${} variables useable here or just hardcode it
-version="${file.jarVersion}" #mandatory
+version="2.0" #mandatory
 # A display name for the mod
 displayName="More Time Capsules Mod" #mandatory
 # A URL to query for updates for this mod. See the JSON update specification <here>
